requirements.txt
flask
google-cloud-storage
google-cloud-vision
celery
google-auth
google-auth-oauthlib
redis